requirements.txt
streamlit>=1.28.0
requests>=2.32.3,<2.33.0
sseclient-py>=1.8.0
python-dateutil>=2.8.2
openai>=1.68.0
pandas>=2.2.0
tabulate>=0.9.0
python-dotenv>=1.0.0 